Tasks/NutritionFacts/requirements.txt
pandas==2.1.2
python-dotenv==1.0.0
tensorflow==2.14.0